html/parsing/manakai/foreign-tags-1.dat

#data
<!DOCTYPE html><svg><filterres></filterRes></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg filterres>

#data
<!DOCTYPE html><svg filterres=""/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       filterres=""
